+++ zigbee-api/src/main/java/org/bubblecloud/zigbee/network/zcl/protocol/ZclDataType.java
@@ -10,16 +10,16 @@ public enum ZclDataType {
     N_X_NEIGHBORS_INFORMATION("N X Neighbors information",Object.class),
     N_X_UNSIGNED_16_BIT_INTEGER("N X Unsigned 16-bit integer",Object.class),
     N_X_UNSIGNED_8_BIT_INTEGER("N X Unsigned 8-bit integer",Object.class),
-    SIGNED_16_BIT_INTEGER("Signed 16-bit integer",Short.class),
-    SIGNED_8_BIT_INTEGER("Signed 8-bit integer",Byte.class),
-    UNSIGNED_16_BIT_INTEGER("Unsigned 16-bit integer",Short.class),
+    SIGNED_16_BIT_INTEGER("Signed 16-bit integer",Integer.class),
+    SIGNED_8_BIT_INTEGER("Signed 8-bit integer",Integer.class),
+    UNSIGNED_16_BIT_INTEGER("Unsigned 16-bit integer",Integer.class),
     UNSIGNED_32_BIT_INTEGER("Unsigned 32-bit integer",Integer.class),
-    UNSIGNED_8_BIT_INTEGER("Unsigned 8-bit integer",Byte.class),
-    _16_BIT_BITMAP("16-bit bitmap",Short.class),
-    _16_BIT_ENUMERATION("16-bit enumeration",Short.class),
-    _8_BIT_BITMAP("8-bit bitmap",Byte.class),
-    _8_BIT_DATA("8-bit data",Byte.class),
-    _8_BIT_ENUMERATION("8-bit enumeration",Byte.class);
+    UNSIGNED_8_BIT_INTEGER("Unsigned 8-bit integer",Integer.class),
+    _16_BIT_BITMAP("16-bit bitmap",Integer.class),
+    _16_BIT_ENUMERATION("16-bit enumeration",Integer.class),
+    _8_BIT_BITMAP("8-bit bitmap",Integer.class),
+    _8_BIT_DATA("8-bit data",Integer.class),
+    _8_BIT_ENUMERATION("8-bit enumeration",Integer.class);
     private final String label;
     private final Class dataClass;
